==> notebooks/02_annotation.ipynb <==
# Imports
import pandas as pd

# Charger les données augmentées
df = pd.read_csv("../data/augmented/podcasts_keywords_augmented.csv")

df.head()

df["is_kid_friendly"] = None
df.head()

import ast

def to_list(x):
    if isinstance(x, list):
        return x
    try:
        return ast.literal_eval(x)   # lit la string comme liste python
    except:
        return []

df["keywords_clean"] = df["keywords_clean"].apply(to_list)
unique_keywords = sorted({kw for lst in df["keywords_clean"] for kw in lst})

print("Nombre de keywords uniques :", len(unique_keywords))
print(unique_keywords)


non_kid_keywords = ['1946 terror',
                    '911 dispatcher stress',
                    'abuse in hollywood',
                    'abuse of power',
                    'abuser capture',
                    'abusive relationship',
                    'addiction',
                    'addiction medicine',
                    'addiction psychology',
                    'addiction recovery',
                    'addiction treatment', 
                    'adult dating life',
                    'adult family life',
                    'adult film actor',
                    'adult humor', 
                    'adult living situation', 
                    'adult relationships', 
                    'adult topics', 
                    'adult work humor', 
                    'adulting challenges', 
                    'adulting struggles', 
                    'afghanistan', 
                    'ageism in film',
                    'alcohol', 
                    'alcohol abuse', 
                    'alcohol recovery', 
                    'alcoholism in fame', 
                    'alpha male', 
                    'antidepressants', 
                    'anxiety', 
                    'anxiety and fame', 
                    'anxiety humor', 
                    'anxiety in fame', 
                    'apocalypse now', 
                    'apology debate', 
                    'assisted suicide', 
                    'atlanta child murders', 
                    'authoritarianism', 
                    'autonomous weapons', 
                    'autopsy', 
                    'autopsy controversy', 
                    'ax murder case', 
                    'axe murder historical', 
                    'ayahuasca', 
                    'ayahuasca journey', 
                    'ayahuasca retreat', 
                    'backlash', 
                    'backstage drama', 
                    'backstage fight', 
                    'backstage secrets', 
                    'bad boss stories', 
                    'bad boy records', 
                    'bad breath', 
                    'bad date confessions', 
                    'bad marriage advice', 
                    'bad sex stories', 
                    'betrayal and deception', 
                    'betrayal of trust', 
                    'betrayal story', 
                    'black market economy', 
                    'breakup recovery', 
                    'breakup tips', 
                    'btk killer',
                    'california crimes', 
                    'capitalism vs socialism', 
                    'celebrity addiction', 
                    'celebrity breakup', 
                    'celebrity burnout', 
                    'celebrity death mystery', 
                    'celebrity divorce', 
                    'celebrity drug use', 
                    'celebrity judge scandal', 
                    'celebrity substance abuse', 
                    'celebrity trial', 
                    'celebrity vulnerability', 
                    'child abuse', 
                    'child abuse cover-up', 
                    'child abuse exoneration', 
                    'child abuse hysteria', 
                    'child abuse lawsuit', 
                    'child abuse warning signs', 
                    'child beauty pageant', 
                    'child brides', 
                    'child custody battle', 
                    'child exploitation', 
                    'child medical abuse', 
                    'child migrant story', 
                    'child neglect charges', 
                    'child stardom trauma', 
                    'childhood trauma', 
                    'cleveland torso murderer',
                    'cold war artifacts', 
                    'cold war espionage', 
                    'cold war history', 
                    'cold war secrets', 
                    'combat mindset', 
                    'community fear', 
                    'controversial dating', 
                    'controversial debate', 
                    'controversial figure', 
                    'convicted killer', 
                    'conviction', 
                    'crime documentaries', 
                    'crime legacy', 
                    'crime of greed', 
                    'crime of passion', 
                    'crime scene', 
                    'crime scene investigation', 
                    'crime solving techniques', 
                    'criminal charges', 
                    'criminal justice issues', 
                    'criminal justice reform', 
                    'criminal psychology', 
                    'cringe', 
                    'cringe stories', 
                    'cruise ship crime', 
                    'cult escape', 
                    'cult member confession', 
                    'cult suicide', 'cults', 
                    'culture war debate', 
                    'culture wars', 
                    'custody battle', 
                    'custody battle aftermath', 
                    'custody battle closure', 
                    'custody battle evidence', 
                    'custody battle ordeal', 
                    'custody battle tactics', 
                    'custody battle weaponization', 
                    'custody case corruption', 
                    'custody case drama', 
                    'custody hearing', 
                    'custody loss', 
                    'custody retaliation', 
                    'custody system failure', 
                    'custody victory', 
                    'custody warfare', 
                    'dark humor', 
                    'dark passingering', 
                    'dark side of fame', 
                    'dateline', 
                    'dating', 
                    'dating advice', 
                    'dating apps', 
                    'dating disasters', 
                    'dating fails', 
                    'dating horror stories', 
                    'dating in the spotlight', 
                    'dating on tv', 
                    'dave chappelle', 
                    'daytime tv therapy', 
                    'deadpool franchise', 
                    'death', 
                    'death acceptance', 
                    'death and grieving', 
                    'death investigation', 
                    'death stranding', 
                    'decade-old evidence', 
                    'declassified documents', 
                    'declassified secrets', 
                    'decline and fall', 
                    'delphi murders', 
                    'deportation', 
                    'deportation fear', 
                    'deportation risk', 
                    'depression', 
                    'depression era crime', 
                    'detective cold case', 
                    'detective drama', 
                    'detective perspective', 
                    'dirty jobs', 
                    'disappearance', 
                    'disappearance in florida', 
                    'disappearance without a trace', 
                    'divorce', 
                    'divorce battle', 
                    'divorce stories', 
                    'divorce war', 
                    'divorce weaponization', 
                    'domestic abuse coverup', 
                    'domestic dispute', 
                    'domestic violence awareness', 
                    'domestic violence case', 
                    'domestic violence charges', 
                    'domestic violence warning',
                    'drinking stories','drug abuse stories', 
                    'drug addiction', 
                    'drug cartel violence', 
                    'drug enforcement', 
                    'drug legalization debate', 
                    'drug policy reform', 
                    'drug relapse', 
                    'drug use as a minor',
                    'emotional abuse', 
                    'emotional aftermath', 
                    'emotional burnout', 
                    'emotional devastation', 
                    'emotional distress',
                    'emotional manipulation',
                    'emotional toll of war', 
                    'emotional toll of work', 
                    'emotional weight',
                    'emotions', 
                    'emptiness', 
                    'end-of-life care', 
                    'end-of-life choices', 
                    'end-of-life process',
                    'epstein', 
                    'epstein victims fund', 
                    'escape from cult', 
                    'escape from manipulation', 
                    'escape from sect', 
                    'espionage history', 
                    'espionage secrets', 
                    'espionage techniques', 
                    'ethical compromise', 
                    'ethical debate', 
                    'ethical dilemma', 
                    'evil and motive', 
                    'existential crisis', 
                    'existential dread', 
                    'existential risk', 
                    'false abuse claims', 
                    'false abuse claims defense', 
                    'false accusation defense', 
                    'false accusations', 
                    'false accusations lawsuit', 
                    'false allegation damage', 
                    'false allegations', 
                    'false allegations defense', 
                    'false flag death', 
                    'false imprisonment', 
                    'false medical claims', 
                    'false medical diagnosis', 
                    'false munchausen allegation', 
                    'false reporting of domestic violence', 
                    'false reports', 
                    'false sexual abuse allegations', 
                    'false sexual abuse claim', 
                    'familial annihilation', 
                    'familial dna analysis', 
                    'familial homicide', 
                    'familial murder', 
                    'family arguments', 
                    'family betrayal', 
                    'family conflict', 
                    'family court', 
                    'family court abuse', 
                    'family court bias', 
                    'family court drama', 
                    'family court errors', 
                    'family court flaws', 
                    'family court injustice', 
                    'family court reform', 
                    'family despair', 
                    'family drama confessions', 
                    'family estrangement', 
                    'family grief',
                    'family separation', 
                    'family separation trauma', 
                    'family struggle', 
                    'family suspicion', 
                    'family trauma', 
                    'famous murder trial', 
                    'fear and isolation', 
                    'fear and survival', 
                    'fear excitement', 
                    'fears', 
                    'femininity',
                    'fertility struggle', 
                    'final words', 
                    'financial anxiety', 
                    'financial crime', 
                    'financial crime network', 
                    'financial dispute', 
                    'financial elites',
                    'financial motive',
                    'financial stress', 
                    'financial struggle', 
                    'fire rescue stories', 
                    'forbidden history', 
                    'forced marriage', 
                    'fox news controversy',
                    'gang violence', 
                    'gary ridgway', 
                    'gender bias in work', 
                    'gender identity', 
                    'gene expression', 
                    'geopolitical tension', 
                    'ghost stories', 
                    'ghosting', 
                    'ghosting experience', 
                    'global conspiracy',
                    'golden state killer', 
                    "grant's death", 
                    'green river killer', 
                    'grief', 
                    'grief after loss', 
                    'grief and loss', 
                    'grief and trauma', 
                    'grieving process', 
                    'guards negligence', 
                    'gun control debate', 
                    'healing childhood trauma', 
                    'healing from trauma', 
                    'healing trauma', 
                    'hidden life of suspect', 
                    'hidden motive', 
                    'high altitude danger', 
                    'high finance corruption', 
                    'high society scandal',
                    'historical scandal', 
                    'historical serial killer', 
                    'historical spy', 
                    'historical tragedy', 
                    'historical trauma', 
                    'historical trial', 
                    'historical true crime', 
                    'hollywood dark side', 
                    'hollywood pressure', 
                    'hollywood scandal', 
                    'hollywood secrets', 
                    'hollywood shame', 
                    'hollywood showrunner', 
                    'hollywood tragedy', 
                    'holocaust', 
                    'home life horror', 
                    'homicide investigation', 
                    'homicide motive', 
                    'horror of the past',
                    'hospital drama',
                    'human exploitation',
                    'human manipulation',
                    'human smuggling', 
                    'human trafficking', 
                    'human trafficking investigation',
                    'husband as suspect', 
                    'illegal activity', 
                    'illegal activity past', 
                    'illegal drug trade',  
                    'immigration enforcement', 
                    'in-laws trouble',
                    'inmate violence', 
                    'inner struggles', 
                    'insider trading scandal',
                    'intellectual dark web', 
                    'internet fame burnout',
                    'investigation', 
                    'investigation dead ends', 
                    'investigation failure',
                    'jeffrey epstein', 
                    'jeffrey epstein associates', 
                    'jeffrey epstein autopsy', 
                    'jeffrey epstein death', 
                    'jeffrey epstein finances', 
                    'jeffrey epstein island', 
                    'jeffrey epstein lawyers', 
                    'jeffrey epstein money', 
                    'jeffrey epstein victims', 
                    "jehovah's witness cult", 
                    'jfk assassination',
                    'journalism in conflict',
                    'judge bias', 
                    'judge disagreements', 
                    'judge feuds', 
                    "judge's behavior", 
                    "judges' comments", 
                    'judicial misconduct', 
                    'jury deliberation', 
                    'just war theory', 
                    'justice for abuse', 
                    'justice for mom', 
                    'justice for victim', 
                    'justice for victims', 
                    'justice or revenge', 
                    'justice reform', 
                    'justice system corruption', 
                    'justice system failure', 
                    'kidnapping case', 
                    'killer robots', 
                    'kitty genovese murder', 
                    'kurt cobain death', 
                    'law enforcement corruption', 
                    'law enforcement failure', 
                    'leaving religion', 'legal battle', 
                    'legal battle stress', 
                    'legal battle trauma', 
                    'legal battles', 
                    'legal struggle', 
                    'legal system bias', 
                    'legal system corruption', 
                    'legal system failure', 
                    'legal system flaw', 
                    'legal system flaws', 
                    'legal system manipulation', 
                    'legal system vindication', 
                    'legal troubles',  
                    'legalization debate', 
                    'lgbtq+', 
                    'lgbtq+ representation', 
                    'lgbtq+ rights', 
                    'life and death decisions', 
                    'long-term trauma', 
                    'los angeles crime history', 
                    'loss', 
                    'lost civilization theories',
                    'male bonding', 
                    'male perspective',
                    'marilyn monroe death',
                    'maritime murder', 
                    'marriage and divorce', 
                    'marriage manipulation',
                    'masculinity', 
                    'mass death', 
                    'medical abuse of child',
                    'men vs women', 
                    'mental health crisis', 
                    'mental health debate', 
                    'mental health in prison', 
                    'mental health struggle', 
                    'mental health struggles', 
                    'mental illness comedy', 
                    'mental illness stigma', 
                    'meredith kercher murder', 
                    'mind control', 
                    'mind control aftermath', 
                    'mind control tactics', 
                    'mind control techniques', 
                    'mind manipulation', 
                    'miscarriage of justice', 
                    'miscarriage story',
                    'modern dating', 
                    'modern love',
                    'money laundering',
                    'morgue', 
                    'mormon faith crisis', 
                    'mormon mission experience', 
                    'mormonism', 
                    'mormonism controversy', 
                    'murder', 
                    'murder acquittal', 
                    'murder confession', 
                    'murder conviction', 
                    'murder for hire', 
                    'murder investigation', 
                    'murder motive', 
                    'murder of celebrity', 
                    'murder plot', 'naked', 
                    'near death experiences', 
                    'new suspect lead', 
                    'new years failure', 
                    'new york city crime',
                    'no body no crime', 
                    'online harassment', 
                    'online negativity', 
                    'parental abuse', 
                    'parental alienation', 
                    'parental alienation awareness', 
                    'parental alienation damage', 
                    'parental alienation recovery', 
                    'parental alienation syndrome', 
                    'parental conflict', 
                    'parental guilt', 
                    'parental negligence', 
                    'parental pressure',
                    'personal grief',
                    'personal risk',
                    'phantom killer', 
                    'philosophical zombies',
                    'poisoning case', 
                    'police brutality', 
                    'police bungling', 
                    'police corruption allegations', 
                    'police disbelief', 
                    'police failure', 
                    'police incompetence', 
                    'police interrogation', 
                    'police interrogation analysis', 
                    'police interrogation tactics', 
                    'police interview tapes', 
                    'police investigation', 
                    'police misconduct', 
                    'police mistakes', 
                    'police pressure', 
                    'police procedural', 
                    'police reform debate', 
                    'police rescue', 
                    'police wiretap', 
                    'political corruption', 
                    'political crisis', 
                    'political scandal',
                    'political violence',
                    'post-show trauma', 
                    'post-war crime', 
                    'post-war thought', 
                    'postpartum anxiety', 
                    'power and corruption', 
                    'predator accountability', 
                    'predator capture', 
                    'predator lair', 
                    'predatory behavior', 
                    'prison break attempt', 
                    'prison corruption', 
                    'prison experience', 
                    'prison gangs', 
                    'prison interview', 
                    'prison system failure', 
                    'prison system flaws', 
                    'prison time', 
                    'professional anxiety', 
                    'professional burnout',
                    'professional stress', 
                    'prohibition era', 
                    'property disputes', 
                    'prosecution arguments', 
                    'psychological abuse',
                    'psychological illness', 
                    'psychological recovery', 
                    'psychological torture', 
                    'psychological trauma', 
                    'racial humor', 
                    'racial injustice', 
                    'racial profiling', 
                    'racial tensions', 
                    'racism', 
                    'reality tv manipulation', 
                    'reality tv pressure', 
                    'reality tv relationships', 
                    'reality tv scandals', 
                    'reality tv secrets', 
                    'reality tv stress', 
                    'red flags in men', 
                    'red pill philosophy',
                    'refugee crisis', 
                    'refugee stories', 
                    'regis philbin feud', 
                    'regrets', 
                    'rehab journey', 
                    'rehab stories', 
                    'relationship addiction', 
                    'relationship failure', 
                    'relationship issues', 
                    'relationship red flags', 
                    'relationship scandal', 
                    'relationship toxicity', 
                    'relationship trauma', 
                    'religious abuse', 
                    'religious extremism', 
                    'religious trauma', 
                    'satanic panic',
                    "scandal and grey's anatomy", 
                    'scandal balls', 
                    'scott peterson trial', 
                    'serial killer', 
                    'serial killer capture', 
                    'serial killer patterns', 
                    'serial killer profile', 
                    'serial killer victim', 
                    'serial murder history', 
                    'serial murderer', 
                    'serial rapist', 
                    'sex addiction', 
                    'sex advice', 
                    'sex offenders', 
                    'sex trafficking', 
                    'sex work debate', 
                    'sexual abuse', 
                    'sexual abuse allegations', 
                    'sexual abuse coverup', 
                    'sexual abuse location', 
                    'sexual abuse of minor', 
                    'sexual abuse survivors', 
                    'sexual abuse victims', 
                    'sexual assault survivor story', 
                    'sexual content in comedy', 
                    'sexual health risks', 
                    'sexual liberation', 
                    'sexual misconduct controversy', 
                    'sexual politics', 
                    'sexual slavery', 
                    'sexual violence crimes', 
                    'sexuality', 
                    'sexualization of women', 
                    'shame and guilt', 
                    'shared living stress', 
                    'shoulder injury', 
                    'show business dark side',
                    'show manipulation', 
                    'show pressure', 
                    'small town horror', 
                    'small town tragedy', 
                    'social dilemma', 
                    'social media crime',
                    'social media scrutiny', 
                    'social media toxicity', 
                    'sports betting', 
                    'sports scandals', 
                    'stigma and shame', 
                    'streaming wars', 
                    'stress and disease', 
                    'stressful job', 
                    'suicide note', 
                    'suicide or murder debate', 
                    'suicide versus murder debate', 
                    'suspect', 
                    'suspect analysis', 
                    'suspect confession', 
                    'suspect interrogation', 
                    'suspect interview breakdown', 
                    'suspect rumors', 
                    'suspect sketches', 
                    'suspect zero', 
                    'suspicious death', 
                    'systemic racism', 
                    'taboo topics', 
                    'ted bundy victims', 
                    'texarkana moonlight murders', 
                    'texas cold case', 
                    'tinder culture', 
                    'tinder dating fails', 
                    'torture and dismemberment', 
                    'toxic abuse', 
                    'toxic cycles', 
                    'toxic ex-husband', 
                    'toxic exes', 
                    'toxic family dynamics', 
                    'toxic friendship', 
                    'toxic marriage', 
                    'toxic masculinity', 
                    'toxic relationships', 
                    'toxic relatives', 
                    'toxic work environment', 
                    'toxic workplace',
                    'train crime', 
                    'transgender controversy', 
                    'transgender experience', 
                    'trauma and recovery', 
                    'trauma and resilience', 
                    'trauma exposure', 
                    'trauma healing', 
                    'trauma surgery', 
                    'trial', 
                    'trial of the century', 
                    'trial preparations', 
                    'true crime', 
                    'true crime analysis', 
                    'true crime breakthroughs', 
                    'true crime details', 
                    'true crime documentaries', 
                    'true crime documentation', 
                    'true crime drama', 
                    'true crime ethics', 
                    'true crime genre', 
                    'true crime history', 
                    'true crime legend', 
                    'true crime mystery', 
                    'true crime obsession', 
                    'true crime public opinion', 
                    'true crime puzzle', 
                    'true crime tragedy', 
                    'unemployment stress', 
                    'unexplained crash', 
                    'unexplained death', 
                    'unexplained disappearance', 
                    'unidentified aerial phenomena',  
                    'unidentified killer', 
                    'unresolved homicide',
                    'unsolved case', 
                    'unsolved crime', 
                    'unsolved crimes', 
                    'unsolved details', 
                    'unsolved disappearance', 
                    'unsolved double homicide', 
                    'unsolved hip hop crime', 
                    'unsolved hollywood mystery', 
                    'unsolved murder', 
                    'unsolved mysteries', 
                    'unsolved mystery', 
                    'unsolved serial murder', 
                    'victim advocacy', 
                    'victim impact', 
                    'victim impact statements', 
                    'victim testimony', 
                    "victim's fear", 
                    'victimhood', 
                    'victimology', 
                    'wedding stress', 
                    'weed', 
                    'whiskey craft', 
                    'work anxiety', 
                    'worst boss stories', 
                    'wrongful accusation', 
                    'wrongful conviction', 
                    'wrongful conviction risk', 
                    'wrongful death', 
                    'wrongful imprisonment',
                    'zodiac killer',
                    'zodiac killer ciphers']
non_kid_keywords = [kw.strip().lower() for kw in non_kid_keywords]
print("Nombre de keywords non appropriés :", len(non_kid_keywords))
print(non_kid_keywords)


def is_row_kid_friendly(keywords, bad_list):
    if not isinstance(keywords, list):
        return 1  # approprié

    keywords = [kw.strip().lower() for kw in keywords]

    # Vérifier si un élément de la ligne est dans la liste non appropriée
    for kw in keywords:
        if kw in bad_list:
            return 0  # non approprié
    
    return 1  # approprié
df["is_kid_friendly"] = df["keywords_clean"].apply(
    lambda x: is_row_kid_friendly(x, non_kid_keywords)
)


print(df["is_kid_friendly"].value_counts())
df.head(100)


output_path = "../data/annotated/podcasts_annotated.csv"

df.to_csv(output_path, index=False, encoding="utf-8")

print("Fichier sauvegardé avec succès :", output_path)
